refactor: extract 2 methods to dataHandlers file
